refactor: move out the code that is in init.py files
